% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtruncLC.R
\name{rtruncLC}
\alias{rtruncLC}
\title{Generate truncated random deviates}
\usage{
rtruncLC(n, spec, a = -Inf, b = Inf, ..., its = F)
}
\arguments{
\item{n}{a positive integer for the number of random deviates generated}

\item{spec}{a character value that specifies the underlying probability distribution
\itemize{
  \item{"binom" - stats' binomial distribution}
  \item{"pois" - stats' Poisson distribution}
  \item{"nbinom" - stats' negative binomial distribution}
  \item{"geom" - stats' geometric distribution}
  \item{"gamma" - stats' gamma distribution}
  \item{"norm" - stats' normal distribution}
  \item{"invgauss" - actuar's inverse-Gaussian distribution}
}}

\item{a}{a numeric value for the lower bound of the random variable}

\item{b}{a numeric value for the upper bound of the random variable}

\item{...}{other arguments are are passed to the corresponding distribution-related functions}

\item{its}{boolean, whether classical quantile-based sampling à-la truncdist should be performed. Default is FALSE, so Devroye sampling is used instead.}
}
\value{
A vector with one or more random deviates. An attribute reports on acceptance rates.
}
\description{
This function generates n random deviates that are drawn from the specified truncated distribution via Devroye sampling method for log-concave distributions. It is meant to improve on truncdist package over a subset a distributions, only exponential family models for now.
}
\details{
This function returns a vector of size n, irrespective of parameters supplied.
\itemize{
 \item{Log-concave distributions are such that their logarithmic transformation is concave. A rejection sampling algorithm ensuring acceptance rate > 1/4 exists for log-concave densities \insertCite{Devroye_1986}{truncLCdist}. Its counterpart for log-concave discrete distributions ensures acceptance rate > 1/5 \insertCite{Devroye_1987}{truncLCdist}.}
 \item{Gamma distribution with shape parameter < 1 sampled is simulated by transforming samples from an exponential power distribution, obtained à-la Devroye. Exponential power distribution is implemented only for internal use for now, because it is not exponential class and we should provide ad hoc checks, distinct from those we meant to focus in this package.}
}
}
\examples{
rtruncLC(10000, "pois", a=15, lambda=1)
rtruncLC(10000, "binom", a=50, size=100, prob=0.15)
}
\references{
\insertAllCited{}
}
\author{
Euloge Clovis Kenne Pagui, \email{kenneeg@gmail.com}

Michele Lambardi di San Miniato, \email{michele.lambardidisanminiato@gmail.com}
}
