% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncLCdistoptions.R
\name{truncLCdistoptions}
\alias{truncLCdistoptions}
\title{Add/modify distributions}
\usage{
truncLCdistoptions(..., simplify = FALSE, update = list())
}
\arguments{
\item{...}{a character string or named argument with value, to read/write the behavior of the package}

\item{simplify}{a boolean, telling whether multiple options should be presented as vectors in read mode}

\item{update}{allows specific values to be updated dynamically rather than via named key=value pairs, can ignore it, see futile.options for more details}
}
\description{
Sets the package behavior with each distribution, which can be added or modified. Supplies rtruncLC with the relevant information on how each distribution shall be sampled, based on mode (.m), probability mass/density function (.d), cumulative distribution function(.p), whether it is continuous or not (.continuous). Derived from futile.options package by \insertCite{futileopts}{truncLCdist}.
}
\examples{
truncLCdistoptions("binom.d")
truncLCdistoptions(binom.d=stats::dbinom)
}
\references{
\insertAllCited{}
}
